requirements.txt
requests==2.28.1
fastapi==0.78.0
python-dotenv==0.20.0
uvicorn==0.18.2
